chore: renamed models -> schemas, added links to schemas, added place name to a template from the linked data
